project/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
atlastk==0.13.5
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
fastapi==0.115.12
flake8==7.2.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
iso8601==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pycodestyle==2.13.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pypika-tortoise==0.6.1
pytest==8.4.0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
